Chatbot_def/requirements.txt
streamlit==1.37.1
faiss-cpu==1.8.0
sentence-transformers==3.0.1
pypdf==4.3.1
ollama>=0.2.0
numpy==1.26.4
python-dotenv==1.0.1
tqdm==4.66.5
